+++ src/data_struct/mod.rs
@@ -1,9 +1,12 @@
+// Data-structures implemented here are voluntarily more exhaustive than they need to be for the
+// sake of Covercrypt.
 #![allow(dead_code)]
+
 mod dictionary;
-pub mod error;
 mod revision_map;
 mod revision_vec;
 
+pub mod error;
 pub use dictionary::Dict;
 pub use revision_map::RevisionMap;
 pub use revision_vec::RevisionVec;
